Revert "Feature/photo preview"
